pyproject.toml
[tool.poetry]
name = "anyeraproject"
version = "0.1.0"
description = ""
authors = ["arturkalimullin <noti1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.3"
django-environ = "^0.10.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^10.0.0"
django-filter = "^23.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
flake8-blind-except = "^0.2.1"
flake8-bugbear = "^23.6.5"
flake8-builtins = "^2.1.0"
flake8-class-attributes-order = "^0.1.3"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.13.0"
flake8-debugger = "^4.1.2"
flake8-functions = "^0.0.8"
flake8-isort = "^6.0.0"
flake8-mutable = "^1.2.0"
flake8-print = "^5.0.0"
flake8-pytest = "^1.4"
flake8-pytest-style = "^1.7.2"
flake8-quotes = "^3.3.2"
flake8-string-format = "^0.3.0"
flake8-variables-names = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
